Address PR review comments
- loss.py: replace torch.tensor(0.0) accumulators with 0.0 scalar
  to avoid CPU/CUDA device-mismatch when residuals live on GPU.
  Affects BellmanEquationLoss FOC term, EulerEquationLoss
  constrained path, and EulerEquationLoss unconstrained path.
- bellman.py (estimate_bellman_foc_residual): treat dv_dc=None
  from torch.autograd.grad(allow_unused=True) as a zero gradient
  rather than raising. This is a legitimate outcome for
  multi-control models where only some controls affect V(s').
- Move fischer_burmeister from bellman.py to utils.py and update
  loss.py imports and tests accordingly.
